+++ NEWS.md
@@ -1,6 +1,7 @@
 DALEX 0.4.9
 ----------------------------------------------------------------
 * updated `titanic_imputed` ([#113](https://github.com/ModelOriented/DALEX/issues/113)).
+* added `weights` to the explainer. Note that not all explanations know how to handle weights ([#118](https://github.com/ModelOriented/DALEX/issues/118)).
 
 DALEX 0.4.8
 ----------------------------------------------------------------